+++ lib/autoscaler/sidekiq.rb
@@ -1,11 +1,11 @@
 require 'autoscaler/sidekiq/client'
-require 'autoscaler/sidekiq/sleep_wait_server'
+require 'autoscaler/sidekiq/monitor_middleware_adapter'
 
 module Autoscaler
   # namespace module for Sidekiq middlewares
   module Sidekiq
     # Sidekiq server middleware
     # Performs scale-down when the queue is empty
-    Server = SleepWaitServer
+    Server = MonitorMiddlewareAdapter
   end
 end